requirements.txt
fastapi
uvicorn
jinja2
python-jose
mysql-connector-python
